requirements.txt
feedparser==6.0.10
chromadb>=1.3.5
sentence-transformers>=5.1.2
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
ollama>=0.6.1
streamlit>=1.28.0
beautifulsoup4>=4.12.0

